pythonProject1/mysite/requirements-production.txt
asgiref==3.7.2
certifi==2023.11.17
Django==5.0
sqlparse==0.4.4
tzdata==2023.3
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
